requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
fastapi
uvicorn
mlflow
